avm2: Move TextLine's validity enum to fte.rs
This places it next to others from FTE.
