+++ core/src/display_object/text_line.rs
@@ -17,34 +17,7 @@ use gc_arena::{Collect, Gc, Mutation};
 use ruffle_common::avm_string::AvmString;
 use ruffle_common::utils::HasPrefixField;
 use ruffle_macros::istr;
-use ruffle_wstr::WStr;
 use std::sync::Arc;
-
-#[derive(Clone, Copy, Collect, PartialEq)]
-#[collect(require_static)]
-pub enum TextLineValidity {
-    Valid,
-    Invalid,
-    Static,
-    PossiblyInvalid,
-    UserInvalid,
-}
-
-impl TextLineValidity {
-    pub fn parse(string: &WStr) -> Self {
-        if string == b"valid" {
-            Self::Valid
-        } else if string == b"invalid" {
-            Self::Invalid
-        } else if string == b"static" {
-            Self::Static
-        } else if string == b"possiblyInvalid" {
-            Self::PossiblyInvalid
-        } else {
-            Self::UserInvalid
-        }
-    }
-}
 
 #[derive(Clone, Collect, Copy)]
 #[collect(no_drop)]
